Created print function

+++ components/common.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 
+USER_ID=$(sudo id -u)
+if [ ${USER_ID} -ne 0 ]; then
+  echo "\e[36m You need to be root user to run the script"
+  exit 2
+fi
+
 LOG_FILE=/tmp/jenkins.log
 rm -rf $LOG_FILE
+
+if []
 Print() {
   echo -e "\e[36m $1 \e[0m"
 }
